+++ src/JsonResponseBody.php
@@ -4,16 +4,12 @@ declare(strict_types=1);
 namespace ResponseInterop\Impl;
 
 use ResponseInterop\Interface\ResponseBodyHandler;
-use ResponseInterop\Interface\ResponseSenderService;
+use ResponseInterop\Interface\ResponseBodySenderService;
 use ResponseInterop\Interface\ResponseStruct;
 
 class JsonResponseBody implements ResponseBodyHandler
 {
-    public const int DEFAULT_FLAGS = JSON_HEX_TAG
-        | JSON_HEX_APOS
-        | JSON_HEX_AMP
-        | JSON_HEX_QUOT
-        | JSON_THROW_ON_ERROR;
+    public const int DEFAULT_FLAGS = JSON_HEX_TAG | JSON_HEX_APOS | JSON_HEX_AMP | JSON_HEX_QUOT | JSON_THROW_ON_ERROR;
 
     /**
      * @param ?non-empty-string $type
@@ -33,23 +29,21 @@ class JsonResponseBody implements ResponseBodyHandler
      */
     public function prepareResponse(ResponseStruct $response) : void
     {
-        $response->headers->setHeader(
-            'content-type',
-            $this->type ?? 'application/json'
-        );
+        $response->headers
+            ->setHeader('content-type', $this->type ?? 'application/json');
     }
 
     /**
      * @inheritdoc
      */
-    public function sendResponseBody(ResponseSenderService $sender) : void
+    public function sendResponseBody(ResponseBodySenderService $bodySender) : void
     {
-        $sender->sendResponseBodyString(
+        $bodySender->sendResponseBodyString(
             (string) json_encode(
                 $this->data,
                 $this->flags ?? static::DEFAULT_FLAGS,
                 $this->depth ?? 512,
-            )
+            ),
         );
     }
 }